feat: 메인페이지 구조 정리
