+++ src/app/home/Body.tsx
@@ -1,114 +1,231 @@
-/* eslint-disable no-unused-vars */
+/* eslint-disable object-curly-newline */
+/* eslint-disable import/no-unresolved */
+/* eslint-disable function-paren-newline */
 /* eslint-disable operator-linebreak */
+/* eslint-disable implicit-arrow-linebreak */
 /* eslint-disable consistent-return */
+/* eslint-disable react-hooks/exhaustive-deps */
 
 'use client';
 
-import { useState } from 'react';
-import { useForm } from 'react-hook-form';
-import Latest from './Latest.tsx';
-import Distance from './Distance.tsx';
+import Script from 'next/script';
+import { Map, MapMarker } from 'react-kakao-maps-sdk';
+import Image from 'next/image';
+import { useEffect, useRef, useState } from 'react';
+import { format } from 'date-fns';
+import { ko } from 'date-fns/locale';
+import { useInfiniteQuery } from '@tanstack/react-query';
+import { searchMeetings } from '@/api/nearMeeting.ts';
+import category from '@/util/category.json';
+import Post from '../components/Post.tsx';
 
-function Body() {
-  const [activeTab, setActiveTab] = useState<string>('latest');
-  const [searchLocation, setSearchLocation] = useState<string>('');
-  const { register: registerSearch, handleSubmit: handleSubmitSearch } =
-    useForm();
+function Distance() {
+  const [latitude, setLatitude] = useState<number | null>(null);
+  const [longitude, setLongitude] = useState<number | null>(null);
+  const [locationFetched, setLocationFetched] = useState(false);
 
-  const handleTabClick = (tabName: string) => {
-    if (activeTab !== tabName) {
-      setActiveTab(tabName);
+  useEffect(() => {
+    function successCallback(position: any) {
+      // 완료 되면 콘솔 지우기!!
+      console.log('Latitude: ', position.coords.latitude);
+      console.log('Longitude: ', position.coords.longitude);
+      setLatitude(position.coords.latitude);
+      setLongitude(position.coords.longitude);
+      setLocationFetched(true); // 위치 정보를 성공적으로 가져온 후 설정
     }
+
+    function errorCallback(error: any) {
+      // 함수 매개변수의 타입 지정
+      console.error('Error Code = ', error.code, ' - ', error.message);
+    }
+    navigator.geolocation.getCurrentPosition(successCallback, errorCallback);
+  }, []);
+
+  const locPosition =
+    latitude && longitude ? { lat: latitude, lng: longitude } : null;
+
+  const renderSize = 8;
+
+  const { data, fetchNextPage, hasNextPage, isFetchingNextPage } =
+    useInfiniteQuery({
+      queryKey: ['Latest Meeting List', latitude, longitude],
+      queryFn: async ({ pageParam = 0 }) => {
+        if (latitude !== null && longitude !== null) {
+          return searchMeetings(
+            latitude,
+            longitude,
+            pageParam,
+            renderSize,
+            'asc',
+          );
+        }
+      },
+      getNextPageParam: (lastPage, allPages) => {
+        if (lastPage === null) {
+          return undefined;
+        }
+        return allPages.length;
+      },
+      enabled: locationFetched, // 위치 정보를 성공적으로 가져왔을 때만 쿼리 실행
+      initialPageParam: 0,
+    });
+
+  const loadMoreRef = useRef<HTMLDivElement | null>(null);
+
+  useEffect(() => {
+    if (!hasNextPage) return;
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries[0].isIntersecting) {
+          fetchNextPage();
+        }
+      },
+      {
+        root: null, // 기본적으로 브라우저 뷰포트를 root로 사용
+        rootMargin: '0px',
+        threshold: 0.1, // 타겟 요소가 10% 보이면 콜백 실행
+      },
+    );
+    if (loadMoreRef.current) observer.observe(loadMoreRef.current);
+
+    return () => {
+      if (loadMoreRef.current) observer.disconnect();
+    };
+  }, [hasNextPage, fetchNextPage, loadMoreRef.current]);
+
+  const [selectedTime, setSelectedTime] = useState<string>('');
+  const [selectedTitle, setSelectedTitle] = useState<string>('');
+  const [selectedAddress, setSelectedAddress] = useState<string>('');
+  const [selectedImg, setSelectedImg] = useState<string>('');
+  const [selectedCategory, setSelectedCategory] = useState<number | null>(null);
+
+  const handleMain = (marker: any) => {
+    setSelectedTime(
+      format(new Date(marker.time), 'PPP EEEE', {
+        locale: ko,
+      }),
+    );
+    setSelectedTitle(marker.mainTitle);
+    setSelectedAddress(marker.address);
+    setSelectedCategory(marker.category_id);
+    setSelectedImg(marker.mainImg);
   };
 
-  const mainBody = () => {
-    if (activeTab === 'latest') {
-      return (
-        <div className="flex w-full items-center justify-center">
-          <Distance />
-        </div>
-      );
-    }
-    if (activeTab === 'distance') {
-      return (
-        <div className="flex w-full items-center justify-center">
-          <Latest />
-        </div>
-      );
-    }
-  };
+  const markers =
+    data?.pages
+      .filter((page) => page && page.data && page.data.meeting_list)
+      .flatMap((page) =>
+        page.data.meeting_list.map(
+          (meeting: {
+            image_url: string;
+            title: string;
+            meeting_id: number;
+            road_address_name: string;
+            latitude: number;
+            longitude: number;
+            appointment_time: string;
+            category_id: number;
+          }) => ({
+            lat: meeting.latitude,
+            lng: meeting.longitude,
+            meetingId: meeting.meeting_id,
+            mainTitle: meeting.title,
+            address: meeting.road_address_name,
+            mainImg: meeting.image_url,
+            time: meeting.appointment_time,
+            category_id: meeting.category_id,
+          }),
+        ),
+      ) || [];
 
   return (
-    <div className="flex h-full w-full flex-col">
-      <div className="mt-10 flex w-full flex-row items-center justify-between p-5 sm:p-5 md:p-2 lg:p-8">
-        {/* 탭 리스트 */}
-        <div
-          role="tablist"
-          className="m tabs-lgsm:w-1/2 tabs tabs-bordered w-[17rem] sm:w-1/2 md:w-1/3 lg:w-1/4"
-        >
-          <input
-            type="radio"
-            name="latestTab"
-            role="tab"
-            className={`tab ${activeTab === 'latest' ? 'font-bold' : ''}`}
-            aria-label="최신순"
-            checked={activeTab === 'latest'}
-            onChange={() => handleTabClick('latest')}
-          />
-
-          <input
-            type="radio"
-            name="distanceTab"
-            role="tab"
-            className={`tab mt-4 ${activeTab === 'distance' ? 'font-bold' : ''}`}
-            aria-label="거리순"
-            checked={activeTab === 'distance'}
-            onChange={() => handleTabClick('distance')}
-          />
+    <div className="mt-[-0.5rem] flex h-full w-[30rem] flex-col sm:w-[30rem] md:w-[43rem] lg:w-[65rem]">
+      <div className="flex flex-row items-center justify-center">
+        <div className="m-4 flex h-[28rem] w-60 flex-col border border-zinc-300">
+          <figure className="relative flex min-h-60 min-w-60">
+            <Image
+              src={selectedImg || category.basic_image[4]}
+              fill
+              alt="Image"
+              sizes="100vm"
+            />
+          </figure>
+          <div className="flex h-full flex-col justify-around p-2 py-8">
+            <p className="text-lg font-bold">{selectedTime}</p>
+            <p className="text-xl">
+              {selectedTitle || (
+                <span className="flex items-center justify-center text-lg font-bold text-zinc-500">
+                  모임을 선택해주세요.
+                </span>
+              )}
+            </p>
+            <p className="text-sm text-zinc-400">{selectedAddress}</p>
+            {/* <p className="text-sm text-zinc-400">({selectedAddressName})</p> */}
+            <p className="text-sm text-zinc-400">
+              {selectedCategory !== null ? `#${selectedCategory}` : ''}
+            </p>
+          </div>
         </div>
-
-        {/* 지역 검색 드롭다운 */}
-        <div className="h-30 flex w-56 flex-row items-center justify-end lg:w-full">
-          {activeTab === 'latest' && (
-            <form
-              className="border-1 flex w-32 flex-row items-center rounded-lg border border-zinc-600"
-              onSubmit={handleSubmitSearch(() => {
-                setSearchLocation('localKeyword');
-              })}
+        <div
+          className="m-4 h-[28rem] w-[50rem] border border-zinc-300"
+          id="map"
+        >
+          <Script
+            src={process.env.NEXT_PUBLIC_KAKAO_SDK_URL}
+            strategy="beforeInteractive"
+          />
+          {locPosition ? (
+            <Map
+              center={locPosition}
+              style={{ width: '100%', height: '100%' }}
+              level={3}
             >
-              <input
-                {...registerSearch('localKeyword', { required: true })}
-                type="text"
-                className="mx-2 h-full w-full focus:outline-none"
-                placeholder="주소 입력"
-                autoComplete="off"
-              />
-              <button type="submit">
-                <svg
-                  className="m-2 h-6 w-6"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeWidth="2"
-                    d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                  />
-                </svg>
-              </button>
-            </form>
+              <MapMarker position={locPosition} />
+              {markers.map((marker, index) => (
+                <MapMarker
+                  key={index}
+                  position={{ lat: marker.lat, lng: marker.lng }}
+                  onClick={() => handleMain(marker)}
+                />
+              ))}
+            </Map>
+          ) : (
+            <p className="flex items-center justify-center">Loading map...</p>
           )}
         </div>
       </div>
-      <div className="flex h-full w-full flex-col items-start justify-start">
-        {mainBody()}
-        {/* 강제로 지도 컴포넌트 렌더링 */}
-        <div className="h-0 w-0 overflow-hidden">
-          <Distance />
-        </div>
+      <div className="my-20 flex flex-row flex-wrap items-center">
+        {data?.pages // pages가 정의된 경우에만 접근
+          .filter((page) => page && page.data && page.data.meeting_list)
+          .flatMap((page) =>
+            page.data.meeting_list.map(
+              (meeting: {
+                image_url: string;
+                title: string;
+                meeting_id: number;
+                road_address_name: string;
+                category_id: number;
+                end_time: string;
+              }) => (
+                <Post
+                  key={meeting.meeting_id}
+                  titleImage={
+                    meeting.image_url ||
+                    category.category_image[meeting.category_id]
+                  }
+                  title={meeting.title}
+                  meetingId={meeting.meeting_id}
+                  location={meeting.road_address_name}
+                  endTime={meeting.end_time}
+                />
+              ),
+            ),
+          )}
+        {isFetchingNextPage && <div>Loading...</div>}
+        <div ref={loadMoreRef} />
       </div>
     </div>
   );
 }
 
-export default Body;
+export default Distance;
