+++ src/app/home/page.tsx
@@ -1,16 +1,14 @@
-// 홈 페이지 컴포넌트
-
 'use server';
 
-import Header from './Banner.tsx';
-import Body from './Body.tsx';
+import Header from './Header.tsx';
+import Distance from './Body.tsx';
 
 async function page() {
   return (
     <div className="flex h-full w-full flex-col items-center justify-center ">
       <Header />
-      <div className="flex h-full w-96 flex-col items-center justify-center sm:w-[30rem] md:w-[46rem] lg:w-[58rem] xl:w-[67.5rem]">
-        <Body />
+      <div className="mt-10 flex w-full flex-row items-center justify-center p-5 sm:p-5 md:p-2 lg:p-8">
+        <Distance />
       </div>
     </div>
   );
