+++ src/util/category.json
@@ -63,6 +63,7 @@
     "https://i.ibb.co/1GDYG85/image.png",
     "https://i.ibb.co/FDSwtTR/image.png",
     "https://i.ibb.co/4N8tqsM/image.png",
-    "https://i.ibb.co/Xtyp45S/image.png"
+    "https://i.ibb.co/Xtyp45S/image.png",
+    "https://i.ibb.co/ZxPwqRL/image.png"
   ]
 }
